requirements_gemini.txt
fastapi
aiofiles
uvicorn
python-multipart
jinja2
python-dotenv
pydantic
requests
google-generativeai
